### 2019-04-23 #1402 @whedon
[PRE REVIEW]: hdme: High-Dimensional Regression with Measurement Error
**Submitting author:** @osorensen (<a href="http://orcid.org/0000-0003-0724-3542">Øystein Sørensen</a>)
**Repository:** <a href="https://github.com/osorensen/hdme" target ="_blank">https://github.com/osorensen/hdme</a>
**Version:** v0.2.3
