### 2019-04-23 #1402 @csoneson
👋Hi @dirmeier, @dmbates, @fabian-s - would you be interested in reviewing this submission for JOSS? It looks like a nice fit to your expertise!
