### 2019-04-24 3361aad
Added doi to all references where available
