+++ paper.bib
@@ -13,16 +13,6 @@ volume = "40",
 year = "2012"
 }
 
-@article{sorensen2015,
- author = {Oystein Sorensen and Arnoldo Frigessi and Magne Thoresen},
- journal = {Statistica Sinica},
- number = {2},
- pages = {809-829},
- publisher = {Institute of Statistical Science, Academia Sinica},
- title = {Measurement error in lasso: Impact and likelihood bias correction},
- volume = {25},
- year = {2015}
-}
 
 
 @book{carroll2006,
@@ -179,7 +169,6 @@ author = "Dirk Eddelbuettel and Conrad Sanderson"
  volume = {25},
  year = {2015}
 }
-
 
 
 
