### 2019-04-24 09e17b4
Encoding fix
