### 2019-04-24 #1404 @osorensen
Thanks to @dirmeier and @dmbates for agreeing to review. Just a note; I added doi to all references in paper.bib and recompiled the article. The content in paper.md is unchanged.
