### 2019-04-25 578ae7a
Updated NEWS.md
