### 2019-04-25 #1404 @osorensen
Thanks for pointing this out, @dmbates. I have now tagged the appropriate commit in the repository to version 0.2.3, which is the one currently on CRAN. Note that there have been some small commits to the development version since the release.

https://github.com/osorensen/hdme/releases
