### 2019-04-29 osorensen/hdme#26 @dirmeier
JOSS submission
Dear Øystein,
I have only very minor revisions for the manuscript itself. I uploaded the PDF with corrections [here](https://www.dropbox.com/s/6d75kdnhimuatgz/10.21105.joss.01404-SD.pdf?dl=0#)

Some other things:
- In your [paper from 2018](https://doi.org/10.1080/10618600.2018.1425626) I believe you describe the GMUS and GMUL already. Correct me if I am wrong, but does that mean you already published the code of these (and their non-generalized Gaussian cases) as code?
- I think it would make sense to cite some at least remotely similar packages like [this](https://cran.r-project.org/web/packages/GLSME/).

Best,
S
