### 2019-05-06 #1404 @dmbates
Not the fault of this package but my initial installation attempt on a clean R install failed because the Rglpk package failed to install.  It has an external dependency on an executable that I needed to track down and get installed before I could install this package.
