### 2019-05-06 #1404 @dmbates
There do not appear to be print methods for many of the model types in the examples that I tried.  If you are going to assign an S3 class to the fitted model then it would be common to provide methods for print and for various extractors like coef.

Also the fitting functions typically take the name of the model type without prepending "fit_" or a similar phrase.  So an "lm" object is created by "lm", not "fit_lm", etc.
