### 2019-05-06 #1404 @dmbates
The tests seem rather perfunctory - for the most part checking dimensions.  A more systematic testing approach would be welcome.
