### 2019-05-06 #1404 @dirmeier
See here for the manuscript's revision itself: https://github.com/osorensen/hdme/issues/26
