### 2019-05-07 osorensen/hdme#26 @osorensen
Dear Simon

Regarding your first point:

Thanks a lot for you comment on the manuscript. Yes, code for running the experiments with GMUS and GMUL were included as supplementary material in the [2018 paper](https://www.tandfonline.com/doi/full/10.1080/10618600.2018.1425626). This supplementary material was not subject to peer review, and when submitting the hdme package to JOSS I considered that paper as a purely methodological paper. The paper also states that code is available in this GitHub repository, but that refers to a much older version of this package, which was not on CRAN.

However, I realize now that this means part of the code in this package is actually part of the supplementary material of an already published paper, and that I should have made you aware of this. That being said, the package currently submitted has much more functionality and methods implemented, documentation and vignette, so much work has been done since then.
