### 2019-05-09 osorensen/hdme#26 @dirmeier
Right, I was just trying to point out if there could be licensing issues with the Journal of Computational and Graphical Statistics.

Other than that I have no comments about the manuscript.  For the code, as Doug Bates pointed out methods like `coef`, `summary`, etc. would be nice to have. The rest of the package/documentation/code is great!
