### 2019-05-09 08d822e
Added doi to Sorensen 2015 paper
