+++ paper.bib
@@ -167,7 +167,8 @@ author = "Dirk Eddelbuettel and Conrad Sanderson"
  publisher = {Institute of Statistical Science, Academia Sinica},
  title = {Measurement error in lasso: Impact and likelihood bias correction},
  volume = {25},
- year = {2015}
+ year = {2015},
+ doi = {doi:10.5705/ss.2013.180}
 }
 
 
