### 2019-05-09 ee4cc6f
Fixed duplicate doi in reference
