+++ paper.bib
@@ -168,7 +168,7 @@ author = "Dirk Eddelbuettel and Conrad Sanderson"
  title = {Measurement error in lasso: Impact and likelihood bias correction},
  volume = {25},
  year = {2015},
- doi = {doi:10.5705/ss.2013.180}
+ doi = {10.5705/ss.2013.180}
 }
 
 
