### 2019-05-09 ba9009b
Updated README with note about GLPK dependency
