### 2019-05-09 #1404 @osorensen
Thanks to both reviewers for the very useful feedback and @csoneson for the summary.

I have now added doi for Sorensen et al. (2015) and written a note about the GLPK dependency in the repository README.

I will work on adding print functions and standardizing names of the fitting functions during the next week.
