### 2019-05-09 de338a3
Trying to install Rglpk on travis ci
