### 2019-05-09 osorensen/hdme#26 @osorensen
Thanks for the clarification, Simon. And thanks for your pdf with corrections. I will work through them now and update the paper.
