+++ paper.bib
@@ -26,7 +26,7 @@ year = "2012"
 
 @book{baesens2014,
  author = {Baesens, Bart},
- title = {Analytics in a Big Data World: The Essential Guide to Data Science and Its Applications},
+ title = {{A}nalytics in a {B}ig {D}ata {W}orld: {T}he {E}ssential {G}uide to {D}ata {S}cience and {I}ts {A}pplications},
  year = {2014},
  isbn = {1118892704, 9781118892701},
  edition = {1st},
@@ -53,8 +53,8 @@ year = {1996}
 @book{buhlmann2011,
   author = {B{\"u}hlmann, Peter and van de Geer, Sara},
   publisher = {Springer, Heidelberg},
-  series = {Springer Series in Statistics},
-  title = {Statistics for high-dimensional data},
+  series = {{S}pringer {S}eries in {S}tatistics},
+  title = {Statistics for {H}igh-{D}imensional {D}ata},
   year = 2011,
   doi = "10.1007/978-3-642-20192-9"
 }
@@ -63,18 +63,17 @@ year = {1996}
 
 
 @article{candes2007,
-author = "Candes, Emmanuel and Tao, Terence",
+author = {Candes, Emmanuel and Tao, Terence},
 doi = "10.1214/009053606000001523",
-fjournal = "The Annals of Statistics",
-journal = "Ann. Statist.",
-month = "12",
-number = "6",
-pages = "2313--2351",
-publisher = "The Institute of Mathematical Statistics",
-title = "The Dantzig selector: Statistical estimation when p is much larger than n",
+journal = {Ann. Statist.},
+month = {12},
+number = {6},
+pages = {2313--2351},
+publisher = {The Institute of Mathematical Statistics},
+title = {The {D}antzig selector: {S}tatistical estimation when p is much larger than n},
 url = "https://doi.org/10.1214/009053606000001523",
-volume = "35",
-year = "2007"
+volume = {35},
+year = {2007}
 }
 
 
