+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: hdme
 Type: Package
 Title: High-Dimensional Regression with Measurement Error
-Version: 0.2.3.9000
+Version: 0.2.3.9001
 Encoding: UTF-8
 Author: Oystein Sorensen
 Maintainer: Oystein Sorensen <oystein.sorensen.1985@gmail.com>
@@ -19,13 +19,12 @@ Imports:
   glmnet (>= 2.0-13), 
   ggplot2 (>= 2.2.1),
   Rdpack, 
-  Rcpp (>= 0.12.15)
+  Rcpp (>= 0.12.15),
+  Rglpk (>= 0.6-1),
 URL: https://github.com/osorensen/hdme
 RdMacros: Rdpack
 Suggests: 
     knitr,
-    Rglpk (>= 0.6-1),
-    lpSolveAPI (>= 5.5.2),
     rmarkdown,
     testthat,
     flare,
