+++ paper.bib
@@ -26,7 +26,7 @@ year = "2012"
 
 @book{baesens2014,
  author = {Baesens, Bart},
- title = {{A}nalytics in a {B}ig {D}ata {W}orld: {T}he {E}ssential {G}uide to {D}ata {S}cience and {I}ts {A}pplications},
+ title = {Analytics in a Big Data World: The Essential Guide to Data Science and Its Applications},
  year = {2014},
  isbn = {1118892704, 9781118892701},
  edition = {1st},
@@ -53,8 +53,8 @@ year = {1996}
 @book{buhlmann2011,
   author = {B{\"u}hlmann, Peter and van de Geer, Sara},
   publisher = {Springer, Heidelberg},
-  series = {{S}pringer {S}eries in {S}tatistics},
-  title = {Statistics for {H}igh-{D}imensional {D}ata},
+  series = {Springer Series in Statistics},
+  title = {Statistics for High-Dimensional Data},
   year = 2011,
   doi = "10.1007/978-3-642-20192-9"
 }
@@ -70,7 +70,7 @@ month = {12},
 number = {6},
 pages = {2313--2351},
 publisher = {The Institute of Mathematical Statistics},
-title = {The {D}antzig selector: {S}tatistical estimation when p is much larger than n},
+title = {The {D}antzig selector: Statistical estimation when p is much larger than n},
 url = "https://doi.org/10.1214/009053606000001523",
 volume = {35},
 year = {2007}
@@ -79,7 +79,7 @@ year = {2007}
 
 @article{james2009,
     author = {James, Gareth M. and Radchenko, Peter},
-    title = "{A generalized Dantzig selector with shrinkage tuning}",
+    title = "{A generalized {D}antzig selector with shrinkage tuning}",
     journal = {Biometrika},
     volume = {96},
     number = {2},
@@ -112,7 +112,7 @@ year = "2010"
 }
 
 @article{sorensen2018,
-  author = {Oystein Sorensen and Kristoffer Herland Hellton and Arnoldo Frigessi and Magne Thoresen},
+  author = {{\O}ystein S{\o}rensen and Kristoffer Herland Hellton and Arnoldo Frigessi and Magne Thoresen},
   title = {Covariate Selection in High-Dimensional Generalized Linear Models With Measurement Error},
   journal = {Journal of Computational and Graphical Statistics},
   volume = {27},
@@ -127,7 +127,7 @@ year = "2010"
 
 
   @Manual{theussl2019,
-    title = {Rglpk: R/GNU Linear Programming Kit Interface},
+    title = {Rglpk: {R}/{GNU} Linear Programming Kit Interface},
     author = {Stefan Theussl and Kurt Hornik},
     year = {2019},
     note = {R package version 0.6-4},
@@ -136,7 +136,7 @@ year = "2010"
 
 
 @article{eddelbuettel2014,
-title = "RcppArmadillo: Accelerating R with high-performance C++ linear algebra",
+title = "RcppArmadillo: Accelerating {R} with high-performance {C}++ linear algebra",
 journal = "Computational Statistics & Data Analysis",
 volume = "71",
 pages = "1054 - 1063",
@@ -149,8 +149,8 @@ author = "Dirk Eddelbuettel and Conrad Sanderson"
 
 @article{kristensen2014,
 	Author = {Kristensen, Vessela N. and Lingj{\ae}rde, Ole Christian and Russnes, Hege G. and Vollan, Hans Kristian M. and Frigessi, Arnoldo and B{\o}rresen-Dale, Anne-Lise},
-	Journal = {Nature Reviews Cancer},
-	Title = {Principles and methods of integrative genomic analyses in cancer},
[…]
