+++ paper.bib
@@ -1,202 +1,0 @@
-@article{loh2012,
-author = "Loh, Po-Ling and Wainwright, Martin J.",
-doi = "10.1214/12-AOS1018",
-fjournal = "The Annals of Statistics",
-journal = "Ann. Statist.",
-month = "06",
-number = "3",
-pages = "1637--1664",
-publisher = "The Institute of Mathematical Statistics",
-title = "High-dimensional regression with noisy and missing data: Provable guarantees with nonconvexity",
-url = "https://doi.org/10.1214/12-AOS1018",
-volume = "40",
-year = "2012"
-}
-
-
-
-@book{carroll2006,
-	Author = {Raymond J. Carroll and David Ruppert and Leonard A. Stefanski and Ciprian M. Crainiceanu},
-	Publisher = {Chapman \& Hall/CRC},
-	Title = {Measurement Error in Nonlinear Models: A Modern Perspective, Second Edition},
-	address = {Boca Raton, FL, USA},
-	Year = {2006}
-	}
-
-
-@book{baesens2014,
- author = {Baesens, Bart},
- title = {Analytics in a Big Data World: The Essential Guide to Data Science and Its Applications},
- year = {2014},
- isbn = {1118892704, 9781118892701},
- edition = {1st},
- publisher = {Wiley Publishing},
-}
-
-
-@article{tibshirani1996,
-author = {Tibshirani, Robert},
-title = {Regression Shrinkage and Selection Via the Lasso},
-journal = {Journal of the Royal Statistical Society: Series B (Methodological)},
-volume = {58},
-number = {1},
-pages = {267-288},
-keywords = {quadratic programming, regression, shrinkage, subset selection},
-doi = {10.1111/j.2517-6161.1996.tb02080.x},
-url = {https://rss.onlinelibrary.wiley.com/doi/abs/10.1111/j.2517-6161.1996.tb02080.x},
-year = {1996}
-}
-
-
-
-
-@book{buhlmann2011,
-  author = {B{\"u}hlmann, Peter and van de Geer, Sara},
-  publisher = {Springer, Heidelberg},
-  series = {Springer Series in Statistics},
-  title = {Statistics for High-Dimensional Data},
-  year = 2011,
-  doi = "10.1007/978-3-642-20192-9"
-}
-
-
-
-
-@article{candes2007,
-author = {Candes, Emmanuel and Tao, Terence},
-doi = "10.1214/009053606000001523",
-journal = {Ann. Statist.},
-month = {12},
-number = {6},
-pages = {2313--2351},
-publisher = {The Institute of Mathematical Statistics},
-title = {The {D}antzig selector: Statistical estimation when p is much larger than n},
-url = "https://doi.org/10.1214/009053606000001523",
-volume = {35},
-year = {2007}
-}
-
-
-@article{james2009,
-    author = {James, Gareth M. and Radchenko, Peter},
-    title = "{A generalized {D}antzig selector with shrinkage tuning}",
-    journal = {Biometrika},
-    volume = {96},
-    number = {2},
-    pages = {323-337},
-    year = {2009},
-    month = {06},
-    issn = {0006-3444},
-    doi = {10.1093/biomet/asp013},
-    url = {https://doi.org/10.1093/biomet/asp013},
-    eprint = {http://oup.prod.sis.lan/biomet/article-pdf/96/2/323/691965/asp013.pdf},
-}
-
-
-
-
-
-@article{rosenbaum2010,
-author = "Rosenbaum, Mathieu and Tsybakov, Alexandre B.",
-doi = "10.1214/10-AOS793",
-fjournal = "The Annals of Statistics",
-journal = "Ann. Statist.",
-month = "10",
-number = "5",
-pages = "2620--2651",
[…]
