### 2019-05-09 c7ac286
Removed prefix fit_ from all functions and files
