+++ R/RcppExports.R
@@ -1,7 +1,7 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-fit_mu_lasso <- function(omega, gamma, W, z, betaInit, activeSet) {
-    .Call(`_hdme_fit_mu_lasso`, omega, gamma, W, z, betaInit, activeSet)
+mu_lasso <- function(omega, gamma, W, z, betaInit, activeSet) {
+    .Call(`_hdme_mu_lasso`, omega, gamma, W, z, betaInit, activeSet)
 }
 
