+++ tests/testthat/test_corrected_lasso.R
@@ -9,9 +9,9 @@ sigmaUU <- diag(x = 0.2, nrow = p, ncol = p) # Measurement error covariance matr
 W <- X + rnorm(n, sd = diag(sigmaUU)) # Measurement matrix (this is the one we observe)
 beta <- c(seq(from = 0.1, to = 1, length.out = 5), rep(0, p-5)) # Coefficient
 y <- X %*% beta + rnorm(n, sd = 1) # Response
-fit <- fit_corrected_lasso(W, y, sigmaUU, family = "gaussian", no_radii = 2) # Run the corrected lasso
+fit <- corrected_lasso(W, y, sigmaUU, family = "gaussian", no_radii = 2) # Run the corrected lasso
 
-test_that("fit_corrected_lasso returns right object", {
+test_that("corrected_lasso returns right object", {
   expect_output(str(fit), "List of 2")
   expect_equal(class(fit), "corrected_lasso")
 })
@@ -32,14 +32,14 @@ test_that("cv_corrected_lasso computes correctly", {
 
 # Now check the logistic version
 y <- rbinom(n, size = 1, prob = hdme:::logit(X %*% c(rep(5, 2), rep(0, p-2))))
-fit <- fit_corrected_lasso(W, y, sigmaUU, family = "binomial", no_radii = 6)
+fit <- corrected_lasso(W, y, sigmaUU, family = "binomial", no_radii = 6)
 
-test_that("fit_corrected_lasso returns right object for logistic regression", {
+test_that("corrected_lasso returns right object for logistic regression", {
   expect_output(str(fit), "List of 2")
   expect_equal(class(fit), "corrected_lasso")
 })
 
-test_that("fit_corrected_lasso has right dimension in beta vectors", {
+test_that("corrected_lasso has right dimension in beta vectors", {
           expect_equal(dim(fit$betaCorr), c(5, 6))
           expect_equal(length(fit$radii), 6L)
           })
