+++ R/plot.gmus.R
@@ -1,4 +1,4 @@
-#' @title Plot the estimates returned by fit_gmus and fit_mus
+#' @title Plot the estimates returned by gmus and mus
 #' @description Plot the number of nonzero coefficients along a range of delta
 #'   values if delta has length larger than 1, or the estimated coefficients if
 #'   delta has length 1.
@@ -20,7 +20,7 @@
 #' # Response
 #' y <- X %*% beta + rnorm(n, sd = 1)
 #' # Run the MU Selector
-#' mus1 <- fit_mus(W, y)
+#' mus1 <- mus(W, y)
 #' # Draw an elbow plot to select delta
 #' plot(mus1)
 #'
@@ -28,7 +28,7 @@
 #' # delta where the curve has an "elbow".
 #' # In this case, the elbow is at about delta = 0.08, so
 #' # we use this to compute the final estimate:
-#' mus2 <- fit_mus(W, y, delta = 0.08)
+#' mus2 <- mus(W, y, delta = 0.08)
 #' # Plot the coefficients
 #' plot(mus2)
 #'

[…]
