### 2019-05-09 c68234e
Started adding S3 methods
