+++ R/gmu_lasso.R
@@ -38,9 +38,9 @@
 #' # Binomial response
 #' y <- rbinom(n, 1, (1 + exp(-X%*%beta))**(-1))
 #' # Run the GMU Lasso
-#' gmu_lasso <- gmu_lasso(W, y, delta = NULL)
+#' fit <- gmu_lasso(W, y, delta = NULL)
 #' # Get an elbow plot, in order to choose delta.
-#' plot(gmu_lasso)
+#' plot(fit)
 #'
 #'
 gmu_lasso <- function(W, y, lambda = NULL, delta = NULL,
