+++ man/print.gds.Rd
@@ -1,0 +1,16 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/print_methods.R
+\name{print.gds}
+\alias{print.gds}
+\title{Print a Generalized Dantzig Selector Object}
+\usage{
+\method{print}{gds}(x, ...)
+}
+\arguments{
+\item{x}{Fitted model object returned by \code{\link{gds}}.}
+
+\item{...}{Other arguments (not used).}
+}
+\description{
+Default print method for a \code{gds} object.
+}

[…]
