### 2019-05-09 0cbd3e8
Finished first iteration of adding S3 methods
