+++ tests/testthat/test_corrected_lasso.R
@@ -1,46 +1,89 @@
-context("Testing corrected lasso")
+# Tests of corrected_lasso
 
 suppressWarnings(RNGversion("3.5.0"))
-set.seed(100)
-n <- 100 # Number of samples
-p <- 5 # Number of covariates
-X <- matrix(rnorm(n * p), nrow = n)  # True (latent) variables
-sigmaUU <- diag(x = 0.2, nrow = p, ncol = p) # Measurement error covariance matrix (typically estimated by replicate measurements)
-W <- X + rnorm(n, sd = diag(sigmaUU)) # Measurement matrix (this is the one we observe)
-beta <- c(seq(from = 0.1, to = 1, length.out = 5), rep(0, p-5)) # Coefficient
-y <- X %*% beta + rnorm(n, sd = 1) # Response
-fit <- corrected_lasso(W, y, sigmaUU, family = "gaussian", no_radii = 2) # Run the corrected lasso
+set.seed(1)
 
-test_that("corrected_lasso returns right object", {
-  expect_output(str(fit), "List of 3")
-  expect_equal(class(fit), "corrected_lasso")
+# Generate example data and create a first fit
+n <- 100
+p <- 50
+X <- matrix(rnorm(n * p), nrow = n)
+sigmaUU <- diag(x = 0.2, nrow = p, ncol = p)
+W <- X + rnorm(n, sd = diag(sigmaUU))
+beta <- c(seq(from = 0.1, to = 1, length.out = 5), rep(0, p-5))
+y <- X %*% beta + rnorm(n, sd = 1)
+fit <- corrected_lasso(W, y, sigmaUU, family = "gaussian")
+
+# First test that the result is as it should
+test_that("corrected_lasso returns correct object", {
+  expect_s3_class(fit, "corrected_lasso")
+  expect_equal(fit$family, "gaussian")
+  expect_equal(dim(fit$betaCorr), c(50, 20))
+  expect_equal(round(fit$betaCorr[3, 5], 7), 0.4869312)
+  expect_equal(round(fit$betaCorr[13, 15], 7), -0.0817454)
+  expect_equal(length(fit$radii), 20)
 })
 
-cv_fit <- cv_corrected_lasso(W, y, sigmaUU, n_folds = 2, no_radii = 10)
-
-test_that("cv_corrected_lasso returns right object", {
-  expect_output(str(cv_fit), "List of 5")
-  expect_equal(class(cv_fit), "cv_corrected_lasso")
+# Next test that it fails when it should
+test_that("corrected_lasso fails when it should", {
+  expect_error(corrected_lasso(W, y))
+  expect_error(corrected_lasso(W, y, sigmaUU, family = "gamma"))
+  expect_error(corrected_lasso(list(W), y, sigmaUU))
+  expect_error(corrected_lasso(W, y, sigmaUU, radii = -1))
+  expect_error(corrected_lasso(W, y, sigmaUU, alpha = 0))
 })
 
-test_that("cv_corrected_lasso computes correctly", {
-  expect_equal(round(cv_fit$radius_min, 2), round(2.19, 2))
-  expect_equal(round(cv_fit$loss_min, 2), round(1.13, 2))
-  expect_equal(round(cv_fit$radius_1se, 2), round(1.10, 2))
-  expect_equal(round(cv_fit$loss_1se, 2), round(1.52, 2))
+# Test that the S3 methods work
+test_that("S3 methods for corrected_lasso work", {
+  expect_output(coef(fit),
+                regexp = "Number of nonzero coefficient estimates")
+  expect_output(print(fit),
+                regexp = "Corrected gaussian lasso object with 50 variables fitted with 20 regularization parameters.")
[…]
