### 2019-05-13 9ce1178
Wrote test methods for GMUS
