### 2019-05-13 4b5c03c
Wrote tests for gmu_lasso
