### 2019-05-13 971b674
Added missing definition of n in gmu_lasso
