### 2019-05-13 ebe38a6
More tests to improve coverage
