+++ tests/testthat/test_corrected_lasso.R
@@ -26,6 +26,10 @@ test_that("corrected_lasso returns correct object", {
 # Next test that it fails when it should
 test_that("corrected_lasso fails when it should", {
   expect_error(corrected_lasso(W, y))
+  expect_error(corrected_lasso(as.character(W), y, sigmaUU))
+  expect_error(corrected_lasso(W, as.character(y), sigmaUU))
+  expect_error(corrected_lasso(W, y, list(sigmaUU)))
+  expect_error(corrected_lasso(W, c(y, 2), sigmaUU, family = "gaussian"))
   expect_error(corrected_lasso(W, y, sigmaUU, family = "gamma"))
   expect_error(corrected_lasso(list(W), y, sigmaUU))
   expect_error(corrected_lasso(W, y, sigmaUU, radii = -1))
@@ -84,6 +88,39 @@ test_that("S3 methods for corrected_lasso work", {
                 regexp = "Number of nonzero coefficient estimates")
   expect_output(print(fit),
                 regexp = "Corrected binomial lasso object with 50 variables fitted with 20 regularization parameters.")
+  expect_s3_class(plot(fit), "ggplot")
+  expect_s3_class(plot(fit, type = "path"), "ggplot")
+})
+
+
+
+# Poisson regression
+suppressWarnings(RNGversion("3.5.0"))
+set.seed(3)
+
+n <- 1000
+p <- 5
+beta <- c(.01, .01, 0, 0, 0)
+X <- matrix(rnorm(n * p), nrow = n)
+sigmaUU <- diag(x = 0.2, nrow = p, ncol = p)
+W <- X + rnorm(n, sd = diag(sigmaUU))
+y <- rpois(n, exp(X %*% beta))
+fit <- corrected_lasso(W, y, sigmaUU, family = "poisson")
+
+# First test that the result is as it should
+test_that("corrected_lasso returns correct object in the poisson case", {
+  expect_s3_class(fit, "corrected_lasso")
+  expect_equal(fit$family, "poisson")
+  expect_equal(dim(fit$betaCorr), c(5, 20))
+  expect_equal(round(fit$betaCorr[3, 5], 6), 0)
+  expect_equal(round(fit$betaCorr[1, 15], 6), 0)
+  expect_equal(length(fit$radii), 20)
+})
+
+# Test that the S3 methods work
+test_that("S3 methods for corrected_lasso work", {
+  expect_output(print(fit),
+                regexp = "Corrected poisson lasso object with 5 variables fitted with 20 regularization parameters.")
   expect_s3_class(plot(fit), "ggplot")
   expect_s3_class(plot(fit, type = "path"), "ggplot")
 })
