### 2019-05-13 495e592
Wrote tests for Poisson regression
