### 2019-05-14 34496b8
Added missing parenthesis in test_plots.R
