### 2019-05-14 6d723ac
Added convergence tests
