+++ R/gmu_lasso.R
@@ -10,6 +10,7 @@
 #'   supported.
 #' @param active_set Logical. Whether or not to use an active set strategy to
 #'   speed up coordinate descent algorithm.
+#' @param maxit Maximum number of iterations of iterative reweighing algorithm.
 #'
 #' @return An object of class "gmu_lasso".
 #' @export
@@ -46,7 +47,7 @@
 #'
 #'
 gmu_lasso <- function(W, y, lambda = NULL, delta = NULL,
-                          family = "binomial", active_set = TRUE){
+                          family = "binomial", active_set = TRUE, maxit = 1000){
 
   if(family == "binomial") {
     mu <- logit
@@ -82,8 +83,6 @@ gmu_lasso <- function(W, y, lambda = NULL, delta = NULL,
   bOld <- stats::rnorm(p)/p
   bNew <- stats::rnorm(p)/p
   IRLSeps <- 1e-7
-  maxit <- 100
-
 
   bhatGMUL <- matrix(nrow=p, ncol=length(delta))
 
@@ -108,12 +107,10 @@ gmu_lasso <- function(W, y, lambda = NULL, delta = NULL,
       count <- count+1
       Diff1 <- sum(abs(bNew - bOld))
       Diff2 <- sum(abs(bNew - bOlder))
-      #print(paste("Diff1 = ", Diff1, ", Diff2 = ", Diff2, sep=""))
     }
-    if(count >= maxit) print(paste("Did not converge"))
+    if(count >= maxit) stop(paste("Did not converge. Consider increasing maxit."))
     bhatGMUL[ ,i] <- bNew
   }
-  ## TODO: Should return a list, including regularization parameters
 
   fit <- list(intercept = bhatGMUL[1, ],
               beta = matrix(bhatGMUL[-1, ] / scales, nrow = p - 1),
