### 2019-05-14 osorensen/hdme#29 @osorensen
Add S3 methods
Add print, summary, and coef methods to the S3 classes defined.
