### 2019-05-14 #1404 @osorensen
Dear @csoneson, @dmbates and @dirmeier

The master branch of the GitHub repository is now updated with development version 0.2.3.9001 of hdme. The following are the most important updates:

- All main methods have been renamed according to the suggestion from @dmbates, i.e., the function names are now `gmus()`, `gmu_lasso()`, .etc., rather than `fit_gmus()`, `fit_gmu_lasso()?, etc.
- print() and coef() methods have been added for all S3 classes, as suggested by @dmbates. Their use is illustrated for the `gmus()` function in [this issue](https://github.com/osorensen/hdme/issues/29), and the examples in the documentation have also been updated to show their use.
- The unit tests have been throughly updated. In addition to testing dimensions and classes, they are now also testing for exact values of estimated coefficients for each of the methods. The test coverage is now 99 % according to the [code coverage report](https://codecov.io/gh/osorensen/hdme?branch=master).
- I have also updated the manuscript according to the comments of @dirmeier in [this issue](https://github.com/osorensen/hdme/issues/26).
