### 2019-05-14 osorensen/hdme#26 @osorensen
A comment on the formatting of references: I have capitalized titles properly in paper.bib, but let the JOSS compiler (i.e., "whedon") decide how the final formatting becomes. This appears to be the recommended way of doing it.
