### 2019-05-14 #1404 @dirmeier
Hello, I think the `fit_gmus` and `fit_corrected_lasso` need to be updated to `gmus` and ` corrected_lasso` in the manuscript. Other than that, I think we are done with the review from my side. Great package!
