### 2019-05-14 ef609ee
Updated function names in paper
