### 2019-05-14 #1404 @osorensen
Thanks @dirmeier, `fit_gmus` and `fit_corrected_lasso` are now updated.
