### 2019-05-15 eb7e7d4
Started with weighting
