+++ R/musalgorithm.R
@@ -7,12 +7,20 @@
 #'
 #' @keywords internal
 #'
-musalgorithm <- function(W, y, lambda, delta){
+musalgorithm <- function(W, y, lambda, delta, weights = NULL){
   # We assume the first column of W is constants, i.e., intercept
   n <- dim(W)[1]
   p <- dim(W)[2]
   obj <- c(rep(1,p),rep(0,p))
   mat <- matrix(0,nrow=4*p, ncol=2*p)
+
+  # Weight matrix
+  if(is.null(weights)){
+    D <- diag(n)
+  } else {
+    D <- diag(weights)
+  }
+
 
   # Inequality constraint, -u_j - beta_j <= 0
   mat[1:p,1:p] <- -diag(p)
@@ -25,18 +33,18 @@ musalgorithm <- function(W, y, lambda, delta){
   # First "score function" constraint
   mat[(2*p+1),1:p] <- matrix(0, nrow=1, ncol=p)
   mat[(2*p+2):(3*p),1:p] <- matrix(-delta, nrow=(p-1), ncol=p)
-  mat[(2*p+1):(3*p),(p+1):(2*p)] <- 1/n*(t(W)%*%W)
+  mat[(2*p+1):(3*p),(p+1):(2*p)] <- 1/n*(t(W) %*% D %*% W)
 
   # Second "score function" constraint
   mat[(3*p+1),1:p] <- matrix(0, nrow=1, ncol=p)
   mat[(3*p+2):(4*p),1:p] <- matrix(-delta, nrow=(p-1), ncol=p)
-  mat[(3*p+1):(4*p),(p+1):(2*p)] <- -1/n*(t(W)%*%W)
+  mat[(3*p+1):(4*p),(p+1):(2*p)] <- -1/n*(t(W) %*% D %*% W)
 
   rhs <- rep(0,(4*p))
-  rhs[(2*p+1)] <- 1/n*(t(W[,1])%*%y)
-  rhs[(2*p+2):(3*p)] <- lambda + 1/n*(t(W[,-1])%*%y)
-  rhs[(3*p+1)] <- -1/n*(t(W[,1])%*%y)
-  rhs[(3*p+2):(4*p)] <- lambda - 1/n*(t(W[,-1])%*%y)
+  rhs[(2*p+1)] <- 1/n*(t(W[,1]) %*% D %*% y)
+  rhs[(2*p+2):(3*p)] <- lambda + 1/n*(t(W[,-1]) %*% D %*% y)
+  rhs[(3*p+1)] <- -1/n*(t(W[,1]) %*% D %*% y)
+  rhs[(3*p+2):(4*p)] <- lambda - 1/n*(t(W[,-1]) %*% D %*% y)
   dir <- rep("<=",4*p)
   bounds <- list(lower=list(ind=1:(2*p), val=rep(-Inf,2*p)),
                  upper=list(ind=1:(2*p), val=rep(Inf,2*p)))
