### 2019-05-15 d76ffe8
Updated tests so they match new output of S3 methods
