### 2019-05-15 7142964
Working case of weighting for GDS
