### 2019-05-15 #1404 @csoneson
@osorensen - thanks for updating the package, and @dirmeier - thanks for letting us know that you are happy with the submission! Although @dmbates has already checked all the boxes in the checklist, I would like to get a confirmation that he is also satisfied with the additions prompted by his comments above, before we move on to the next stage.
