### 2019-05-15 #1404 @dmbates
The modified version has addressed my concerns.
