### 2019-05-15 #1404 @csoneson
@osorensen - I had another look through your paper, and there are a few small things that should be fixed before acceptance:
* A typo in the first paragraph: satisifaction -> satisfaction
* The Eddelbuettel & Sanderson reference is missing the journal name
* For both Sørensen et al papers, there is no first name initial for the first author
* I would suggest being consistent with the use of journal abbreviations (either for all references, or for none of them)

In addition, please tag a release in your GitHub repository, generate a corresponding Zenodo archive, and report the Zenodo DOI back here. The title and authors of the Zenodo archive should match those of the paper.
