### 2019-05-15 65e9045
Started fixing some final paper issues
