+++ paper.bib
@@ -1,8 +1,7 @@
 @article{loh2012,
 author = "Loh, Po-Ling and Wainwright, Martin J.",
 doi = "10.1214/12-AOS1018",
-fjournal = "The Annals of Statistics",
-journal = "Ann. Statist.",
+journal = "The Annals of Statistics",
 month = "06",
 number = "3",
 pages = "1637--1664",
@@ -65,7 +64,7 @@ year = {1996}
 @article{candes2007,
 author = {Candes, Emmanuel and Tao, Terence},
 doi = "10.1214/009053606000001523",
-journal = {Ann. Statist.},
+journal = {The Annals of Statistics},
 month = {12},
 number = {6},
 pages = {2313--2351},
@@ -99,8 +98,7 @@ year = {2007}
 @article{rosenbaum2010,
 author = "Rosenbaum, Mathieu and Tsybakov, Alexandre B.",
 doi = "10.1214/10-AOS793",
-fjournal = "The Annals of Statistics",
-journal = "Ann. Statist.",
+journal = "The Annals of Statistics",
 month = "10",
 number = "5",
 pages = "2620--2651",
@@ -136,15 +134,14 @@ year = "2010"
 
 
 @article{eddelbuettel2014,
-title = "RcppArmadillo: Accelerating {R} with high-performance {C}++ linear algebra",
-journal = "Computational Statistics & Data Analysis",
-volume = "71",
-pages = "1054 - 1063",
-year = "2014",
-issn = "0167-9473",
-doi = "10.1016/j.csda.2013.02.005",
-url = "http://www.sciencedirect.com/science/article/pii/S0167947313000492",
-author = "Dirk Eddelbuettel and Conrad Sanderson"
+  title = {RcppArmadillo: Accelerating {R} with High-Performance {C}++ Linear Algebra},
+  author = {Dirk Eddelbuettel and Conrad Sanderson},
+  journal = {Computational Statistics and Data Analysis},
+  volume = {71},
+  pages = {1054 - 1063},
+  year = {2014},
+  doi = {10.1016/j.csda.2013.02.005},
+  url = {http://www.sciencedirect.com/science/article/pii/S0167947313000492}
 }
 
 @article{kristensen2014,
@@ -169,12 +166,6 @@ author = "Dirk Eddelbuettel and Conrad Sanderson"
  year = {2015},
  doi = {10.5705/ss.2013.180}
 }
-
-
-
-
-
-
 
 
 
