### 2019-05-16 50086c6
Fixed more ms issues
