+++ paper.bib
@@ -5,7 +5,6 @@ journal = "The Annals of Statistics",
 month = "06",
 number = "3",
 pages = "1637--1664",
-publisher = "The Institute of Mathematical Statistics",
 title = "High-dimensional regression with noisy and missing data: Provable guarantees with nonconvexity",
 url = "https://doi.org/10.1214/12-AOS1018",
 volume = "40",
@@ -68,7 +67,6 @@ journal = {The Annals of Statistics},
 month = {12},
 number = {6},
 pages = {2313--2351},
-publisher = {The Institute of Mathematical Statistics},
 title = {The {D}antzig selector: Statistical estimation when p is much larger than n},
 url = "https://doi.org/10.1214/009053606000001523",
 volume = {35},
@@ -102,7 +100,6 @@ journal = "The Annals of Statistics",
 month = "10",
 number = "5",
 pages = "2620--2651",
-publisher = "The Institute of Mathematical Statistics",
 title = "Sparse recovery under matrix uncertainty",
 url = "https://doi.org/10.1214/10-AOS793",
 volume = "38",
@@ -110,14 +107,13 @@ year = "2010"
 }
 
 @article{sorensen2018,
-  author = {{\O}ystein S{\o}rensen and Kristoffer Herland Hellton and Arnoldo Frigessi and Magne Thoresen},
+  author = {Sørensen, Øystein and Hellton, Kristoffer Herland and Frigessi, Arnoldo and Thoresen, Magne},
   title = {Covariate Selection in High-Dimensional Generalized Linear Models With Measurement Error},
   journal = {Journal of Computational and Graphical Statistics},
   volume = {27},
   number = {4},
   pages = {739-749},
   year  = {2018},
-  publisher = {Taylor & Francis},
   doi = {10.1080/10618600.2018.1425626},
   URL = {https://doi.org/10.1080/10618600.2018.1425626},
   eprint = {https://doi.org/10.1080/10618600.2018.1425626}
@@ -156,11 +152,10 @@ year = "2010"
 
 
 @article{sorensen2015,
- author = {{\O}ystein S{\o}rensen and Arnoldo Frigessi and Magne Thoresen},
+ author = {Sørensen, Øystein and Frigessi, Arnoldo and Thoresen, Magne},
  journal = {Statistica Sinica},
  number = {2},
  pages = {809-829},
- publisher = {Institute of Statistical Science, Academia Sinica},
  title = {Measurement error in lasso: Impact and likelihood bias correction},
  volume = {25},
  year = {2015},
