### 2019-05-18 #1404 @osorensen
@csoneson - I have now fixed the points above. For the fourth point, I chose to consistently use full journal names. I have generated a release, with version number 0.3.0, and it has Zenodo DOI  10.5281/zenodo.2939896.
