+++ README.md
@@ -1,21 +1,15 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-
-# hdme
+hdme
+====
 
 <!-- badges: start -->
+[![CRAN\_Status\_Badge](http://www.r-pkg.org/badges/version/hdme)](https://cran.r-project.org/package=hdme) [![Build Status](https://travis-ci.org/osorensen/hdme.svg?branch=master)](https://travis-ci.org/osorensen/hdme) [![Codecov test coverage](https://codecov.io/gh/osorensen/hdme/branch/master/graph/badge.svg)](https://codecov.io/gh/osorensen/hdme?branch=master) <!-- badges: end -->
 
-[![CRAN\_Status\_Badge](http://www.r-pkg.org/badges/version/hdme)](https://cran.r-project.org/package=hdme)
-[![Build
-Status](https://travis-ci.org/osorensen/hdme.svg?branch=master)](https://travis-ci.org/osorensen/hdme)
-[![Codecov test
-coverage](https://codecov.io/gh/osorensen/hdme/branch/master/graph/badge.svg)](https://codecov.io/gh/osorensen/hdme?branch=master)
-<!-- badges: end -->
+The goal of hdme is to provide penalized regression methods for High-Dimensional Measurement Error problems (errors-in-variables).
 
-The goal of hdme is to provide penalized regression methods for
-High-Dimensional Measurement Error problems (errors-in-variables).
-
-## Installation
+Installation
+------------
 
 Install `hdme` from CRAN using.
 
@@ -32,10 +26,7 @@ devtools::install_github("osorensen/hdme", build_vignettes = TRUE)
 
 ### Dependency on Rglpk
 
-`hdme` uses the [Rglpk
-package](https://cran.r-project.org/package=Rglpk), which requires the
-GLPK library package to be installed. On some platforms this requires a
-manual installation.
+`hdme` uses the [Rglpk package](https://cran.r-project.org/package=Rglpk), which requires the GLPK library package to be installed. On some platforms this requires a manual installation.
 
 On Debian/Ubuntu, you might use:
 
@@ -49,72 +40,34 @@ On macOS, you might use:
 brew install glpk
 ```
 
-## Methods
+Methods
+-------
 
 hdme provides implementations of the following algorithms:
 
 The methods implemented in the package include
 
-  - Corrected Lasso for Linear Models (Loh and Wainwright (2012))
-  - Corrected Lasso for Generalized Linear Models (Sorensen, Frigessi,
-    and Thoresen (2015))
-  - Matrix Uncertainty Selector for Linear Models (Rosenbaum and
-    Tsybakov (2010))
-  - Matrix Uncertainty Selector for Generalized Linear Models (Sorensen
-    et al. (2018))
-  - Matrix Uncertainty Lasso for Generalized Linear Models (Sorensen et
-    al. (2018))
-  - Generalized Dantzig Selector (James and Radchenko (2009))
+-   Corrected Lasso for Linear Models (Loh and Wainwright (2012))
+-   Corrected Lasso for Generalized Linear Models (Sorensen, Frigessi, and Thoresen (2015))
+-   Matrix Uncertainty Selector for Linear Models (Rosenbaum and Tsybakov (2010))
+-   Matrix Uncertainty Selector for Generalized Linear Models (Sorensen et al. (2018))
+-   Matrix Uncertainty Lasso for Generalized Linear Models (Sorensen et al. (2018))
+-   Generalized Dantzig Selector (James and Radchenko (2009))
 
-## Contributions
+Contributions
[…]
